==> modelling/01_modelling_prep_target_encoding.ipynb <==
# 01 Modelling Prep (Target Encoding of Cities)

In this notebook we target encode the city names.

from sklearn.model_selection import train_test_split
import pandas as pd

# load old train data
df = pd.read_csv("data/df_train_2.csv", index_col=0)

# split the train data into smaller train set and a set used for target encoding
df_train, df_test = train_test_split(df, test_size=.1, random_state=0)

# load all data to get names of the cities
df_all = pd.read_csv("data/df_all.csv", index_col=0)

# get df for target encode including city names
df_target_encode = df_all.loc[df_test.index]

# get df from old train set that includes city names
tp = df_all.loc[df.index]

# add rows of cities that were not included in the dataset yet
df_target_encode = pd.concat([df_target_encode, tp[tp.city.isin(["Arlington", "Darien", "Plainville"])]], axis=0)

# remove the rows that were added to df_target_encode
df_train = df_train.drop(index=(tp[tp.city.isin(["Arlington", "Darien", "Plainville"])].index))

encoded_target = df_target_encode.groupby(by=["city", "county"])["price"].mean().to_frame().reset_index()

encoded_target.to_csv("data/df_encoded_target.csv")

df_train.to_csv("data/df_train_encoded_target.csv")
