==> modelling/02_08_modelling_final_performance.ipynb <==
# 02 Modelling (Final Performance)

In this notebook, we test our model architecture on unseen data (on the test set).

import pandas as pd
import numpy as np
from sklearn.metrics import mean_squared_error, r2_score, mean_absolute_error, mean_absolute_percentage_error

from xgboost import XGBRegressor
from sklearn.decomposition import PCA
from sklearn.cluster import KMeans
from sklearn.preprocessing import StandardScaler
from sklearn.neighbors import KNeighborsClassifier

df_train= pd.read_csv("data/after_pca_clustering/df_train.csv", index_col=0)
df_test = pd.read_csv("data/after_pca_clustering/df_test.csv", index_col=0)
df_val = pd.read_csv("data/after_pca_clustering/df_val.csv", index_col=0)

X_train = df_train.drop(columns=["price"])
X_test = df_test.drop(columns=["price"])
X_val = df_val.drop(columns=["price"])

y_train = df_train.loc[:, "price"]
y_test = df_test.loc[:, "price"]
y_val = df_val.loc[:, "price"]

X_train = pd.concat([X_train, X_val])

y_train = pd.concat([y_train, y_val])

# some loss functions
def rmse(y_true, y_pred):
    return np.sqrt(np.mean((y_pred - y_true) ** 2))

def smape(y_true, y_pred):
    return np.mean(200 * np.abs(y_pred - y_true) / (np.abs(y_true) + np.abs(y_pred)))

def rmspe(y_true, y_pred):
    epsilon = 1e-6  # Small constant to avoid division by zero
    return np.sqrt(np.mean(((y_true - y_pred) / (y_true + epsilon)) ** 2))

parameters = {"objective": 'reg:squarederror',
            "n_estimators":350,  # Number of boosting rounds
            "learning_rate":0.1,  # Step size shrinkage
            "max_depth":10,  # Maximum depth of a tree
            }

    
model = XGBRegressor(**parameters)
model.fit(X_train, y_train)
y_hat = model.predict(X_test)

print(f"MSE: {mean_squared_error(y_true=y_test, y_pred=y_hat)}")
print(f"MAE: {mean_absolute_error(y_true=y_test, y_pred=y_hat)}")
print(f"MAPE: {mean_absolute_percentage_error(y_true=y_test, y_pred=y_hat)}")
print(f"R2: {r2_score(y_true=y_test, y_pred=y_hat)}")
print(f"RMSE: {rmse(y_true=y_test, y_pred=y_hat)}")
print(f"SMAPE: {smape(y_true=y_test, y_pred=y_hat)}")
print(f"RMSPE: {rmspe(y_true=y_test, y_pred=y_hat)}")    

### Clustering of Residual Error

residual = model.predict(X_train) - y_train

columns = ['longitude', 'latitude', 'county_Fairfax', 'county_Fairfield',
       'county_Hartford', 'county_Litchfield', 'county_Middlesex',
       'county_New Haven', 'county_New London', 'county_Tolland',
       'county_Windham', 'state_Connecticut', 'state_Virginia',
       'cond_desc_Average', 'cond_desc_Average Plus', 'cond_desc_Fair',
       'cond_desc_Good', 'cond_desc_Poor']
columns = list(set(X_train.columns) & set(columns))
X_train_2 = X_train.rename(columns={0:"c_0", 1:"c_1", 2:"c_2", 3:"c_3", 4:"c_4",5:"c_5"})
X_test_2 = X_test.rename(columns={0:"c_0", 1:"c_1", 2:"c_2", 3:"c_3", 4:"c_4",5:"c_5"})



scaler = StandardScaler()

X_train_scaled = X_train_2.drop(columns=columns)
X_test_scaled = X_test_2.drop(columns=columns)
X_train_scaled = pd.DataFrame(scaler.fit_transform(X_train_scaled), columns=X_train_scaled.columns, index=X_train_scaled.index)
X_test_scaled = pd.DataFrame(scaler.transform(X_test_scaled), columns=X_test_scaled.columns, index=X_test_scaled.index)
X_train_scaled = pd.concat([X_train.loc[:, columns], X_train_scaled], axis=1)
X_test_scaled = pd.concat([X_test.loc[:, columns], X_test_scaled], axis=1)

scaler = StandardScaler()
scaled_residuals = scaler.fit_transform(residual.to_frame())

X_train_scaled["residuals"] = scaled_residuals

k = 4 # You can choose the number of clusters based on your analysis
kmeans = KMeans(n_clusters=k, random_state=42)
clusters = kmeans.fit_predict(X_train_scaled)

X_train_scaled = X_train_scaled.drop(columns="residuals")

model = KNeighborsClassifier(n_neighbors=6)
model.fit(X_train_scaled, clusters)
test_labels = model.predict(X_test_scaled)

X_train_3 = X_train_2.copy()
X_test_3 = X_test_2.copy()

X_train_3["cluster"] = clusters
X_test_3["cluster"] = test_labels

X_train_cl0 = X_train_3[X_train_3.cluster == 0].drop(columns="cluster")
X_test_cl0 = X_test_3[X_test_3.cluster == 0].drop(columns="cluster")

y_train_cl0 = y_train[X_train_cl0.index]
y_test_cl0 = y_test[X_test_cl0.index]

X_train_cl1 = X_train_3[X_train_3.cluster == 1].drop(columns="cluster")
X_test_cl1 = X_test_3[X_test_3.cluster == 1].drop(columns="cluster")

y_train_cl1 = y_train[X_train_cl1.index]
y_test_cl1 = y_test[X_test_cl1.index]


X_train_cl2 = X_train_3[X_train_3.cluster == 2].drop(columns="cluster")
X_test_cl2 = X_test_3[X_test_3.cluster == 2].drop(columns="cluster")

y_train_cl2 = y_train[X_train_cl2.index]
y_test_cl2 = y_test[X_test_cl2.index]

X_train_cl3 = X_train_3[X_train_3.cluster == 3].drop(columns="cluster")
X_test_cl3 = X_test_3[X_test_3.cluster == 3].drop(columns="cluster")

y_train_cl3 = y_train[X_train_cl3.index]
y_test_cl3 = y_test[X_test_cl3.index]

parameters = {"objective": 'reg:squarederror',
            "n_estimators":350,  # Number of boosting rounds
            "learning_rate":0.11,  # Step size shrinkage
            "max_depth":10,  # Maximum depth of a tree
            }

    
model = XGBRegressor(**parameters)
model.fit(X_train_cl0, y_train_cl0)
y_hat_0 = model.predict(X_test_cl0) 

print(f"MSE: {mean_squared_error(y_true=y_test_cl0, y_pred=y_hat_0)}")
print(f"MAE: {mean_absolute_error(y_true=y_test_cl0, y_pred=y_hat_0)}")
print(f"MAPE: {mean_absolute_percentage_error(y_true=y_test_cl0, y_pred=y_hat_0)}")
print(f"R2: {r2_score(y_true=y_test_cl0, y_pred=y_hat_0)}")
print(f"RMSE: {rmse(y_true=y_test_cl0, y_pred=y_hat_0)}")
print(f"SMAPE: {smape(y_true=y_test_cl0, y_pred=y_hat_0)}")
print(f"RMSPE: {rmspe(y_true=y_test_cl0, y_pred=y_hat_0)}")    

parameters = {"objective": 'reg:squarederror',
            "n_estimators":350,  # Number of boosting rounds
            "learning_rate":0.1,  # Step size shrinkage
            "max_depth":10,  # Maximum depth of a tree
            }

    
model = XGBRegressor(**parameters)
model.fit(X_train_cl1, y_train_cl1)
y_hat_1 = model.predict(X_test_cl1) 

print(f"MSE: {mean_squared_error(y_true=y_test_cl1, y_pred=y_hat_1)}")
print(f"MAE: {mean_absolute_error(y_true=y_test_cl1, y_pred=y_hat_1)}")
print(f"MAPE: {mean_absolute_percentage_error(y_true=y_test_cl1, y_pred=y_hat_1)}")
print(f"R2: {r2_score(y_true=y_test_cl1, y_pred=y_hat_1)}")
print(f"RMSE: {rmse(y_true=y_test_cl1, y_pred=y_hat_1)}")
print(f"SMAPE: {smape(y_true=y_test_cl1, y_pred=y_hat_1)}")
print(f"RMSPE: {rmspe(y_true=y_test_cl1, y_pred=y_hat_1)}")    

parameters = {"objective": 'reg:squarederror',
            "n_estimators":400,  # Number of boosting rounds
            "learning_rate":0.1,  # Step size shrinkage
            "max_depth":7,  # Maximum depth of a tree
            }

    
model = XGBRegressor(**parameters)
model.fit(X_train_cl2, y_train_cl2)
y_hat_2 = model.predict(X_test_cl2) 

print(f"MSE: {mean_squared_error(y_true=y_test_cl2, y_pred=y_hat_2)}")
print(f"MAE: {mean_absolute_error(y_true=y_test_cl2, y_pred=y_hat_2)}")
print(f"MAPE: {mean_absolute_percentage_error(y_true=y_test_cl2, y_pred=y_hat_2)}")
print(f"R2: {r2_score(y_true=y_test_cl2, y_pred=y_hat_2)}")
print(f"RMSE: {rmse(y_true=y_test_cl2, y_pred=y_hat_2)}")
print(f"SMAPE: {smape(y_true=y_test_cl2, y_pred=y_hat_2)}")
print(f"RMSPE: {rmspe(y_true=y_test_cl2, y_pred=y_hat_2)}")    

parameters = {"objective": 'reg:squarederror',
            "n_estimators":400,  # Number of boosting rounds
            "learning_rate":0.1,  # Step size shrinkage
            "max_depth":5,  # Maximum depth of a tree
            }

    
model = XGBRegressor(**parameters)
model.fit(X_train_cl3, y_train_cl3)
y_hat_3 = model.predict(X_test_cl3) 

print(f"MSE: {mean_squared_error(y_true=y_test_cl3, y_pred=y_hat_3)}")
print(f"MAE: {mean_absolute_error(y_true=y_test_cl3, y_pred=y_hat_3)}")
print(f"MAPE: {mean_absolute_percentage_error(y_true=y_test_cl3, y_pred=y_hat_3)}")
print(f"R2: {r2_score(y_true=y_test_cl3, y_pred=y_hat_3)}")
print(f"RMSE: {rmse(y_true=y_test_cl3, y_pred=y_hat_3)}")
print(f"SMAPE: {smape(y_true=y_test_cl3, y_pred=y_hat_3)}")
print(f"RMSPE: {rmspe(y_true=y_test_cl3, y_pred=y_hat_3)}")    

y_hat_cl = np.concatenate([y_hat_0, y_hat_1, y_hat_2, y_hat_3])
y_test_cl = np.concatenate([y_test_cl0, y_test_cl1, y_test_cl2, y_test_cl3])

print(f"MSE: {mean_squared_error(y_true=y_test_cl, y_pred=y_hat_cl)}")
print(f"MAE: {mean_absolute_error(y_true=y_test_cl, y_pred=y_hat_cl)}")
print(f"MAPE: {mean_absolute_percentage_error(y_true=y_test_cl, y_pred=y_hat_cl)}")
print(f"R2: {r2_score(y_true=y_test_cl, y_pred=y_hat_cl)}")
print(f"RMSE: {rmse(y_true=y_test_cl, y_pred=y_hat_cl)}")
print(f"SMAPE: {smape(y_true=y_test_cl, y_pred=y_hat_cl)}")
print(f"RMSPE: {rmspe(y_true=y_test_cl, y_pred=y_hat_cl)}")    

## Test Regular Model

df_train = pd.read_csv("data/df_train_2.csv", index_col=0)
df_test = pd.read_csv("data/df_test_2.csv", index_col=0)
df_val = pd.read_csv("data/df_val_2.csv", index_col=0)
X_train = df_train.drop(columns=["price", "id"])
X_val = df_val.drop(columns=["price", "id"])
X_test = df_test.drop(columns=["price", "id"])

y_train = df_train.loc[:, "price"]
y_val = df_val.loc[:, "price"]
y_test = df_test.loc[:, "price"]

X_train = pd.concat([X_train, X_val])
y_train = pd.concat([y_train, y_val])

parameters = {"objective": 'reg:squarederror',
            "n_estimators":350,  # Number of boosting rounds
            "learning_rate":0.1,  # Step size shrinkage
            "max_depth":10,  # Maximum depth of a tree
            }

    
model = XGBRegressor(**parameters)
model.fit(X_train, y_train)
y_hat = model.predict(X_test)

print(f"MSE: {mean_squared_error(y_true=y_test, y_pred=y_hat)}")
print(f"MAE: {mean_absolute_error(y_true=y_test, y_pred=y_hat)}")
print(f"MAPE: {mean_absolute_percentage_error(y_true=y_test, y_pred=y_hat)}")
print(f"R2: {r2_score(y_true=y_test, y_pred=y_hat)}")
print(f"RMSE: {rmse(y_true=y_test, y_pred=y_hat)}")
print(f"SMAPE: {smape(y_true=y_test, y_pred=y_hat)}")
print(f"RMSPE: {rmspe(y_true=y_test, y_pred=y_hat)}")
